Add files via upload
This is the initial commit of a proof of concept for a code scramble/descramble game. I hadn't done any drag-and-drop in JS or rotated things with CSS, and wanted to see what that was like. The initial "database" is just a separate PHP file containing the code snippets. Obviously, that should eventually be replaced with some other storage, like maybe reading a directory or a database.